opendingux: Add filesystem skeleton for GCW0
This is a squashed version of all the development
done in the 2012.11 branch.

Auto-mount SCSI disks (for QEMU)
Problems were reported with the SD card emulation of QEMU for MIPS,
so as a workaround we want to be able to mount harddisk images too.

Added network info for eth0 (for QEMU)

Root login on ttyS0

Mount SD cards with UTF8 support enabled

Mount FAT partitions with shorter write flush delay
This reduces the chance of corruption of the external SD card if it is
removed without unmounting it.
Note that this is a FAT-specific mount option.

Don't create /usr/local/home/apps dir
gmenu2x doesn't look there, it looks for apps/ in the root of the
partition, so having an apps dir in the home dir is just confusing.

Run serial getty on ttyS2
UART2 is the one that is exposed via pads on the PCB.

Map common devices to user-friendly mount point names
Developers who want to access the system partition regularly can
create the configuration file /etc/local/mount_system to let it be
mounted on boot. File content doesn't matter: only the presence of
the file is checked.

Use tabs in fstab
Also removed redundant "defaults" for a fs that had other options.

Mount debugfs on boot
We don't enable debugfs by default in our kernel, but if it's enabled
it is convenient to automatically mount it.

system_info: List active network interfaces
This should help people who have trouble memorizing the Zero's IP
address and are not on a system where DNS-SD works.

Call unlockvt before each time the menu respawns

fstab: Add /dev/shm, tweak tmpfs sizes and permissions
Some applications need /dev/shm and while the directly is created by
mdev, the size of /dev is way too small to store useful amounts of
data. So use a dedicated tmpfs mount for /dev/shm.
The specified sizes for /media and /mnt were set to 1 byte. Although
tmpfs rounds this to 4K (page size), I don't want to rely on that,
so I changed it to be 4K explicitly.
I also put a tighter limit on /var size and denied non-root writing to
that directory.

Run the DHCP server only when usb0 is active

Lower priority of the "watchdog" and "swap" init scripts
The "swap" script can run at the very last moment, chances are low
that swap will be needed as soon as the OS starts.
The "watchdog" script must run early; it has been delayed to run
after udev, if we switch to it in the future, as udev would be
responsible for creating /dev/watchdog.

Create /dev/[null, zero, console] when /dev is mounted
Otherwise, the programs started after /dev is mounted but before
the hotplug handler starts to populate it can behave incorrectly.

Add /run as a symlink to /var/run
This is necessary to get udev to work correctly.

Use udev instead of mdev as the hotplug handler

Use xinetd instead of inetd

Mount /dev/mtp as functionfs

Make the MTP daemon run only when the USB cable is plugged

Add configuration scripts for wlan interfaces
Provided the file /etc/local/network/config-wlan[0-9].conf is present,
running "ifup wlan[0-9]" will configure the wireless interface (set
the ESSID, encryption method and passphrase), and obtain an IP address
automatically, according to the config options.
Running "ifdown wlan[0-9]" will de-configure the interface, and kill
the corresponding wpa_supplicant and udhcpc daemons.

Add a rule for udev to automount block devices

Mount /dev as devtmpfs and let udev fill it

Removed useless init script S20mice.sh
It has been obsoleted since the move to udev.

Generate shared keys for Dropbear at init if needed

Remove the tvtool / tvout_config utilities
Those are specific to the A320 version of OpenDingux.

Use /usr/local/etc instead of /etc/local
The former is distribution-agnostic and should be used instead.

Put shadow password file in /usr/local/etc

Disable root password in /etc/passwd

Require password for telnet login

Mount mmcblk0p1 as /media/$(LABEL) if it has one

Add configuration for vsftpd
The busybox ftpd only supports anonymous access.

Add a ftphome directory used as the top-level dir for FTP sessions

Allow anonymous FTP if empty root password is set
A wrapper script checks the root password and if empty, it will pass
an additional config file to vsftpd which enables anonymous access.
The wrapper also allows adding a local config file by the user.

Exec GMenu2X instead or running it inside a shell

Add scripts to blink the power LED when WiFi is enabled

Show the welcoming cow on SSH sessions too
This funny greeter was maybe my very first addition to OpenDingux,
back in 2009. It's a piece of history. Seriously.

Hardcode default values in wlan scripts
This is done to simplify the configuration files.

Added "ap" tool to enable/disable AP mode

Change metric of interface usb0
When both usb0 and wlan0 are present on the system, it has been
reported that both having the same metric creates a huge packet
drop on the wireless interface.

Modify wlan scripts to use rfkill

Use a /30 netmask for usb0
We will never connect more than two devices with this interface.

Do not run ntpdate at boot
We don't have a working network at boot.
Perhaps we should also delay starting of the ntp daemon until a network
is brought up.