opendingux: Added mini device table
This is based on the mini device table from buildroot 2010.11, but with
all commented out devices removed from the table.

In buildroot 2011.08 the permission info was split off to a separate file.
The permission info we need is the same as the default, so I am not
introducing an OpenDingux specific file for that.