opendingux: Copied mini device table from buildroot 2010.11.
This file was removed in buildroot 2011.02 because it was unused there,
but we are still using it.
In buildroot 2011.08 the permission info was split off to a separate file.
The permission info we need is the same as the default, so I am not
introducing an OpenDingux specific file for that.