Revert "ncurses: Build with wide characters if supported by the toolchain"
This reverts commit b996f9bbadceafde484c1b1f1fd29f46cec5fb6f.

When wide char support is enabled, a lot of files change name from
"ncurses" to "ncursesw". In particular, the "<lib>-config" file being
patched by buildroot and the dynamic libs being installed. This breaks
a clean build and also breaks binary compatibility for applications
using ncurses.

The solution might be to install 8-bit char and wide char versions of
ncurses side-by-side. That requires some nontrivial changes to the
build; in the mean time I'm reverting the change.