fs/opendingux: fstab: Add /dev/shm, tweak tmpfs sizes and permissions
Some applications need /dev/shm and while the directly is created by
mdev, the size of /dev is way too small to store useful amounts of
data. So use a dedicated tmpfs mount for /dev/shm.

The specified sizes for /media and /mnt were set to 1 byte. Although
tmpfs rounds this to 4K (page size), I don't want to rely on that,
so I changed it to be 4K explicitly.

I also put a tighter limit on /var size and denied non-root writing to
that directory.