opendingux: Lower priority of the "watchdog" and "swap" init scripts
The "swap" script can run at the very last moment, chances are low
that swap will be needed as soon as the OS starts.

The "watchdog" script must run early; it has been delayed to run
after udev, if we switch to it in the future, as udev would be
responsible for creating /dev/watchdog.