uClibc: Fix dlsym RTLD_NEXT
This is necessary for apitrace to work properly with games (such as
d2x-rebirth) that load the EGL library with dlopen. Without this change there
is no way to interpose `dlopen`.

With this change the apitrace patch can be simplified, as no longer the ANDROID
path needs to be forced to locate EGL/GLES symbols.

The patch comes from
http://lists.uclibc.org/pipermail/uclibc/2013-January/047373.html and
has already been merged upstream. It is thus to be expected it will no
longer be needed for a next uClibc version.