od-network-config: Added package
This allows the user to set a password for logging in. It requires the
use of a shadow password database.

The implementation is in this repo. If it grows larger we might have to
create a dedicated repo for it.

The code was written by me, the icon was drawn by hi-ban.