sdl: Fix patch to prevent FB restore if not cleared on exit
The patch from commit 489c3454 would not restore any of the video state
on exit, which in the case of double buffering means that the panning
could be left to display the second page. If the next application is
not aware of panning, it will draw to the first page and display the
second page, which leads to a black screen.

The fix is to set the panning to (0, 0) explicitly on exit.