sdl: Add a patch to fix keyboard input when not launched from shell
When not started from a shell, SDL was segfaulting while trying to
detach the process from the controlling terminal.

However, it doesn't make any sense for an SDL app to detach from
its controlling terminal, unless the app is a daemon. Since we
have never seen a SDL app detach from its controlling terminal,
even when starting it from a shell, the feature was probably broken.
Thus, the code has been dropped.