Revert "wget: Added a patch to fix build with Gettext 0.18"
This reverts commit 4c1c93c4194ef4d845d7397c04a31ab54ca6219a.

This is not a fix. It is not even a workaround, since it breaks the
build on the two systems I tried it on.

The purpose of the version number is to check whether "po.m4" and
"po/Makefile.in.in" originate from the same gettext release. Just
increasing the version number without replacing the rest of the file
is asking for problems. Also, updating "po/Makefile.in.in" without
updating "m4/po.m4" will ensure they are out of sync. The only way
this can build at all is if the wrong "po.m4" is picked up. So we
have to fix that instead.