sdl: Added a patch to allow for custom joystick names
When not using the old Linux input event interface, SDL parses the
name of a device to know what it's capable of. It expects the
name to match the following string: "Analog %d-axis %d-button %d-hat".

This patch allows to use devices with a custom name.
For instance: "My Joystick Device (Analog 2-axis 6-button 2-hat)"