oprofile: Fix build with GCC 4.9.0
This problem manifests as a compile error:
  undefined reference to `bfd_elf64_powerpc_vec'

For some reason the problem occurs on mipsel-linux-uclibc when
compiling with GCC 4.9.0 but not when compiling with GCC 4.8.2.

I backported a series of upstream patches that fix the issue. The last
patch contains the actual fix, but the other two are needed to make it
apply unmodified.