sdl: Enhance the triple-buffering patch
Previously, it was not possible to change the resolution or pixel
mode with a new call of SDL_SetVideoMode, a black screen was displayed
instead.

The patch was also more complex than needed, with the use of two mutexes
instead of only one.

Finally, a race could occur if FB_FlipHWSurface was called before the
thread was started; this new version of the patch also fixes this.